+++ packages/@react-native-windows/cli/src/e2etest/autolink.test.ts
@@ -5,8 +5,14 @@
  */
 
 import path from 'path';
+import {commanderNameToOptionName} from '@react-native-windows/telemetry';
+
 import {projectConfigWindows} from '../config/projectConfig';
-import {AutolinkWindows, autolinkOptions} from '../runWindows/utils/autolink';
+import {
+  AutolinkWindows,
+  autolinkOptions,
+  AutoLinkOptions,
+} from '../runWindows/utils/autolink';
 import {DOMParser} from '@xmldom/xmldom';
 import {ensureWinUI3Project} from './projectConfig.utils';
 
@@ -283,18 +289,18 @@ test('ensureXAMLDialect - useWinUI3=true in react-native.config.js, useWinUI3=fa
     '<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003"><PropertyGroup><UseWinUI3>true</UseWinUI3></PropertyGroup></Project>';
   expect(al.experimentalFeaturesProps).toEqual(expectedExperimentalFeatures);
 
-    // example packages.config: 
-    // <packages>
-    //   <package id="SuperPkg" version="42"/>
-    //   <package id="Microsoft.ProjectReunion.WinUI" version="0.8" targetFramework="native"/>
-    // </packages>
-    //
-    expect(al.packagesConfig).toContain('Microsoft.ProjectReunion.WinUI');
-    expect(al.packagesConfig).toContain('<package id="SuperPkg" version="42"/>');
-    expect(al.packagesConfig).not.toContain('Microsoft.UI.Xaml');
-  
-    done();
-  });
+  // example packages.config:
+  // <packages>
+  //   <package id="SuperPkg" version="42"/>
+  //   <package id="Microsoft.ProjectReunion.WinUI" version="0.8" targetFramework="native"/>
+  // </packages>
+  //
+  expect(al.packagesConfig).toContain('Microsoft.ProjectReunion.WinUI');
+  expect(al.packagesConfig).toContain('<package id="SuperPkg" version="42"/>');
+  expect(al.packagesConfig).not.toContain('Microsoft.UI.Xaml');
+
+  done();
+});
 
 test('ensureXAMLDialect - useWinUI3=false in react-native.config.js, useWinUI3=true in ExperimentalFeatures.props', async done => {
   const folder = path.resolve('src/e2etest/projects/WithWinUI3');
@@ -320,18 +326,18 @@ test('ensureXAMLDialect - useWinUI3=false in react-native.config.js, useWinUI3=t
     '<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003"><PropertyGroup><UseWinUI3>false</UseWinUI3></PropertyGroup></Project>';
   expect(al.experimentalFeaturesProps).toEqual(expectedExperimentalFeatures);
 
-    // example packages.config: 
-    // <packages>
-    //   <package id="SuperPkg" version="42"/>
-    //   <package id="Microsoft.ProjectReunion.WinUI" version="0.8" targetFramework="native"/>
-    // </packages>
-    //
-    expect(al.packagesConfig).not.toContain('Microsoft.ProjectReunion.WinUI');
-    expect(al.packagesConfig).toContain('<package id="SuperPkg" version="42"/>');
-    expect(al.packagesConfig).toContain('Microsoft.UI.Xaml');
-  
-    done();
-  });
+  // example packages.config:
+  // <packages>
+  //   <package id="SuperPkg" version="42"/>
+  //   <package id="Microsoft.ProjectReunion.WinUI" version="0.8" targetFramework="native"/>
+  // </packages>
+  //
+  expect(al.packagesConfig).not.toContain('Microsoft.ProjectReunion.WinUI');
+  expect(al.packagesConfig).toContain('<package id="SuperPkg" version="42"/>');
+  expect(al.packagesConfig).toContain('Microsoft.UI.Xaml');
+
+  done();
+});
 
 test('ensureXAMLDialect - useWinUI3 not in react-native.config.js, useWinUI3=true in ExperimentalFeatures.props', async done => {
   const folder = path.resolve('src/e2etest/projects/WithWinUI3');
@@ -357,80 +363,6 @@ test('ensureXAMLDialect - useWinUI3 not in react-native.config.js, useWinUI3=tru
     '<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003"><PropertyGroup><UseWinUI3>true</UseWinUI3></PropertyGroup></Project>';
   expect(al.experimentalFeaturesProps).toEqual(expectedExperimentalFeatures);
 
-<<<<<<< HEAD
-    // example packages.config: 
-    // <packages>
-    //   <package id="SuperPkg" version="42"/>
-    //   <package id="Microsoft.ProjectReunion.WinUI" version="0.8" targetFramework="native"/>
-    // </packages>
-    //
-    expect(al.packagesConfig).toContain('Microsoft.ProjectReunion.WinUI');
-    expect(al.packagesConfig).toContain('<package id="SuperPkg" version="42"/>');
-    expect(al.packagesConfig).not.toContain('Microsoft.UI.Xaml');
-  
-    done();
-  });
-  
-  test('ensureXAMLDialect - useWinUI3 not in react-native.config.js, useWinUI3=false in ExperimentalFeatures.props', async done => {
-    const folder = path.resolve('src/e2etest/projects/WithWinUI3');
-    const rnc = require(path.join(folder, 'react-native.config.js'));
-  
-    const config = projectConfigWindows(folder, rnc.project.windows)!;
-    delete config.useWinUI3;
-    const al = new AutolinkTest(
-      {windows: config},
-      {},
-      {
-        check: false,
-        logging: false,
-      },
-    );
-    al.experimentalFeaturesProps = `<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003"><PropertyGroup><UseWinUI3>false</UseWinUI3></PropertyGroup></Project>`;
-    al.packagesConfig = `<packages><package id="SuperPkg" version="42"/><package id="Microsoft.ProjectReunion.WinUI"/></packages>`;
-  
-    const exd = await al.ensureXAMLDialect();
-    expect(exd).toBeTruthy();
-  
-    const expectedExperimentalFeatures = '<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003"><PropertyGroup><UseWinUI3>false</UseWinUI3></PropertyGroup></Project>';
-    expect(al.experimentalFeaturesProps).toEqual(expectedExperimentalFeatures);
-
-    // example packages.config: 
-    // <packages>
-    //   <package id="SuperPkg" version="42"/>
-    //   <package id="Microsoft.ProjectReunion.WinUI" version="0.8" targetFramework="native"/>
-    // </packages>
-    //
-    expect(al.packagesConfig).not.toContain('Microsoft.ProjectReunion.WinUI');
-    expect(al.packagesConfig).toContain('<package id="SuperPkg" version="42"/>');
-    expect(al.packagesConfig).toContain('Microsoft.UI.Xaml');
-  
-    done();
-  });
-  
-  test('Indirect autolink dependency', () => {
-    const autolink = new AutolinkTest(
-      {windows: {folder: __dirname, sourceDir: '.', solutionFile: 'foo.sln'}},
-      {
-        superModule: {
-          name: 'superModule',
-          root: 'theRoot',
-          platforms: {
-            windows: {
-              sourceDir: __dirname,
-              projects: [
-                {
-                  directDependency: true,
-                  projectFile: 'superModule.vcxproj',
-                  cppHeaders: ['Garfield.h', 'Snoopy.h'],
-                  cppPackageProviders: ['FamousAnimalCartoons'],
-                },
-                {
-                  directDependency: false,
-                  projectFile: 'indirect.vcxproj',
-                }
-              ],
-            },
-=======
   // example packages.config:
   // <packages>
   //   <package id="SuperPkg" version="42"/>
@@ -580,23 +512,73 @@ test('Indirect autolink dependency', () => {
                 projectFile: 'indirect.vcxproj',
               },
             ],
->>>>>>> main
           },
           assets: [],
           hooks: {},
           params: [],
         },
       },
-      {
-        check: true,
-        logging: false,
-        proj: 'projects/WithIndirectDependency/windows/WithIndirectDependency/WithIndirectDependency.vcxproj',
-      },
-    );
-    const replacements = autolink.getCppReplacements();
-    expect(replacements.cppIncludes).toMatch(/#include <Garfield.h>/);
-    expect(replacements.cppIncludes).toMatch(/#include <Snoopy.h>/);
-    expect(replacements.cppPackageProviders).toContain(
-      'packageProviders.Append(winrt::FamousAnimalCartoons())',
-    );
-  });
+    },
+    {
+      check: true,
+      logging: false,
+      proj:
+        'projects/WithIndirectDependency/windows/WithIndirectDependency/WithIndirectDependency.vcxproj',
+    },
+  );
+  const replacements = autolink.getCppReplacements();
+  expect(replacements.cppIncludes).toMatch(/#include <Garfield.h>/);
+  expect(replacements.cppIncludes).toMatch(/#include <Snoopy.h>/);
+  expect(replacements.cppPackageProviders).toContain(
+    'packageProviders.Append(winrt::FamousAnimalCartoons())',
+  );
+});
+
+function validateOptionName(
+  name: string,
+  optionName: keyof AutoLinkOptions,
+): boolean {
+  // Do not add a default case here. Every item must explicitly return true
+  switch (optionName) {
+    case 'check':
+    case 'logging':
+    case 'sln':
+    case 'proj':
+    case 'telemetry':
+      return true;
+  }
+  throw new Error(
+    `Unable to find ${optionName} to match '${name}' in AutoLinkOptions.`,
+  );
+}
+
+test('autolinkOptions - validate options', () => {
+  for (const commandOption of autolinkOptions) {
+    // Validate names
+    expect(commandOption.name).not.toBeNull();
+    expect(commandOption.name.startsWith('--')).toBe(true);
+    expect(commandOption.name).toBe(commandOption.name.trim());
+
+    // Validate defaults
+    if (
+      !commandOption.name.endsWith(' [string]') &&
+      !commandOption.name.endsWith(' [number]')
+    ) {
+      // Commander ignores defaults for flags, so leave undefined to prevent confusion
+      expect(commandOption.default).toBeUndefined();
+    }
+
+    // Validate description
+    expect(commandOption.description).not.toBeNull();
+    expect(commandOption.description!).toBe(commandOption.description!.trim());
+
+    // Validate all command options are present in RunWindowsOptions
+    const optionName = commanderNameToOptionName(commandOption.name);
+    expect(
+      validateOptionName(
+        commandOption.name,
+        optionName as keyof AutoLinkOptions,
+      ),
+    ).toBe(true);
+  }
+});