+++ packages/@rnw-scripts/integrate-rn/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @rnw-scripts/integrate-rn
 
-This log was last generated on Tue, 26 Oct 2021 05:06:56 GMT and should not be manually modified.
+This log was last generated on Thu, 18 Nov 2021 06:11:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.1.1
+
+Thu, 18 Nov 2021 06:11:21 GMT
+
+### Patches
+
+- Use '@react-native-windows/fs' (ngerlem@microsoft.com)
+- Bump @react-native-windows/find-repo-root to v0.0.0-canary.26
+- Bump @react-native-windows/fs to v1.0.0
+- Bump @react-native-windows/package-utils to v0.0.0-canary.23
+- Bump react-native-platform-override to v1.6.1
+- Bump @rnw-scripts/eslint-config to v1.1.9
 
 ## 1.1.0