+++ packages/react-native-platform-override/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - react-native-platform-override
 
-This log was last generated on Tue, 26 Oct 2021 05:06:56 GMT and should not be manually modified.
+This log was last generated on Thu, 18 Nov 2021 06:11:19 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.6.1
+
+Thu, 18 Nov 2021 06:11:19 GMT
+
+### Patches
+
+- Use '@react-native-windows/fs' (ngerlem@microsoft.com)
+- Bump @react-native-windows/fs to v1.0.0
+- Bump @react-native-windows/package-utils to v0.0.0-canary.23
+- Bump @rnw-scripts/eslint-config to v1.1.9
 
 ## 1.6.0